initial commit of cms
this is version 3 of my simple cms. the focus of this version is to
include the database as an sqlite file rather than requiring a user to
create their own mysql database . along with this change, i'm trying to
improve usability and simplify the whole experience of building a
website using this cms.